Misc i64 to u32 change for max_size_allow pow

+++ src/main.rs
@@ -102,7 +102,7 @@ fn setup_logging(verbosity: u64, chain: bool) -> Result<()> {
 async fn main() -> Result<()> {
     let chain = true;
     let verbosity: u64 = 1; //matches.occurrences_of("verbose");
-    let max_size_allow: i64 = (5_f32 * (i64::pow(1024, 3) as f32)) as i64;
+    let max_size_allow: i64 = (5_f32 * (u32::pow(1024, 3) as f32)) as i64;
     let data_dir = util::get_data_dir("", "", PROGRAM_NAME)?;
     util::create_data_dir(&data_dir)?;
